requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.2.0
propcache==0.3.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
sib-api-v3-sdk==7.6.0
six==1.17.0
SQLAlchemy==2.0.40
twilio==9.5.1
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
